requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
typing-extensions>=4.5.0
openai>=1.0.0
websockets>=12.0
httpx>=0.25.0
pydantic>=2.0.0
